Update Source.cpp
After a few days of spotty internet I am now able to upload this. Now the program reads a file, and then posts the puzzle to a new file. Right now I am working on an algorithm to solve the puzzle so that there is an answer key for the people that can't solve it